PciBusDxe: Separation of the PCI device registration and start
BZ: https://bugzilla.tianocore.org/show_bug.cgi?id=2194

The separation of the PCI device registration phase includes only the
installation of the PCI IO Protocol on the PCI node to acquire the EFI
handles, and loading of its applicable PCI Option ROM.

The separation of the PCI device start phase only includes the code that
enables the PCI Bridge device as a Bus Master.

This code change is made in order to introduce the enabling of the other
PCI features in the PCI Bus driver.

Signed-off-by: Ashraf Javeed <ashraf.javeed@intel.com>
Cc: Jian J Wang <jian.j.wang@intel.com>
Cc: Hao A Wu <hao.a.wu@intel.com>
Cc: Ray Ni <ray.ni@intel.com>