PciBusDxe: New PCI Express feature Completion Timeout
BZ: https://bugzilla.tianocore.org/show_bug.cgi?id=2313

The code changes are made; as per the PCI Express Base Specification 4
Revision 1; to enable the configuration of PCI Express feature Completion
Timeout (CTO), that enables the PCI function to wait on programmed dura-
tion for its transactions before timeout, or disable its detection mecha-
nism.

The code changes are made to configure only those PCI devices which are
requested by platform for override through the new PCI Express Platform
protocol interface for device-specific policies. The changes are made to
also comply with the device-specific capability attributes.

The code follows the below implementation specific rules in case the req-
uested platform policy does not match with the device-specific capability
attributes:-
(1) if device is capable of Range A only and if platform ask for any of
    ranges B, C, D; than this implementation will only program the default
    range value for the duration of 50us to 50ms
(2) if device is capable of Range B, or range B & C, or Ranges B, C & D
    only and if the platform ask for the Range A; than this implementation
    will only program the default range value for the duration of 50us to
    50ms
(3) if the device is capable of Range B only, or the ranges A & B; and if
    the platform ask for Range C, or Range D values, than this implement-
    ation will only program the Range B value for the duration of 65ms to
    210ms
(4) if the device is capable of Ranges B & C, or Ranges A, B, and C; and
    if the platform ask for Range D values; than this implementation will
    only program the Range C for the duration of 1s to 3.5s

Signed-off-by: Ashraf Javeed <ashraf.javeed@intel.com>
Cc: Jian J Wang <jian.j.wang@intel.com>
Cc: Hao A Wu <hao.a.wu@intel.com>
Cc: Ray Ni <ray.ni@intel.com>